requirements.txt
# Dépendances du serveur Flask
Flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
Werkzeug==2.3.7
Jinja2==3.1.2

# Base de données
mysql-connector-python==8.1.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Traitement des données
numpy==1.26.0
pandas==2.1.1

# Communication et réseau
requests==2.31.0
websockets==11.0.3

# Dépendances pour le Raspberry Pi
pyserial==3.5
smbus2==0.4.2
pynmea2==1.19.0
adafruit-circuitpython-adxl345==1.14.5
adafruit-circuitpython-dht==3.7.8
RPi.GPIO==0.7.1

# Utilitaires
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3

